+++ src/source/onetouchtv.ts
@@ -536,7 +536,7 @@ export class OnetouchtvScrapper extends BaseProvider {
             streamRow.duration = (info.hours * 60 + info.minutes).toString();
           }
           if (playlist) {
-            streamRow.playlist = playlist;
+            // streamRow.playlist = playlist;
             streamRow.hash = hashSHA256(playlist);
           }
           streamRows.push(streamRow);