+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.9.2 2026-06-03
+
+- Add: Proxy for kisskh encrypted subtitle
+- Add: Wrap flaresolverr to get cookies bypass for kisskh
+- Fix: subtitle content-type
+
 # v0.9.1 2026-05-30
 
 - Add: ntfy have priority